amaranth 0.5.2: dsp cores + vectorscope + sim work
